refactor: util 테스트 패키지 생성 및 이동
