+++ src/test/kotlin/lotto/util/LottoGeneratorTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 import lotto.util.LottoGenerator
 import org.assertj.core.api.Assertions.*
